Deleted video and fixed animation imports?
