+++ website/_ext/video.py
@@ -1,209 +1,169 @@
 """
-Video extention to embed video in a html sphinx output.
-Kept as a reference, but animation.py is the one intended to be actually used.
+Local nodes which may be used to create videos.
+Useful links:
+
+
+rst directive writing documentation:
+https://docutils.sourceforge.io/docs/howto/rst-directives.html
+
+image directive source code:
+https://github.com/docutils/docutils/blob/master/docutils/docutils/parsers/rst/directives/images.py
 """
 
-from pathlib import Path
-from typing import Any, Dict, List, Tuple
-from urllib.parse import urlparse
+from typing import List
+
+import pathlib
 
 from docutils import nodes
-from docutils.parsers.rst import directives
-from sphinx.application import Sphinx
-from sphinx.environment import BuildEnvironment
-from sphinx.util import logging
-from sphinx.util.docutils import SphinxDirective, SphinxTranslator
-
-logger = logging.getLogger(__name__)
-
-SUPPORTED_MIME_TYPES: Dict[str, str] = {
-    ".mp4": "video/mp4",
-    ".ogm": "video/ogg",
-    ".ogv": "video/ogg",
-    ".ogg": "video/ogg",
-    ".webm": "video/webm",
-}
-"Supported mime types of the link tag"
-
-SUPPORTED_OPTIONS: List[str] = [
-    "autoplay",
-    "controls",
-    "height",
-    "loop",
-    "muted",
-    "poster",
-    "preload",
-    "width",
-]
-"List of the supported options attributes"
+from sphinx.util import docutils
+from sphinx.writers import html as html_writers
+from sphinx import application, environment
+from sphinx.builders import html as html_builders
+from sphinx.environment import collectors
 
 
-def get_video(src: str, env: BuildEnvironment) -> Tuple[str, str]:
+class source(nodes.Inline, nodes.Element):
     """
-    Return video and suffix.
-    Load the video to the static directory if necessary and process the suffix. Raise a warning if not supported but do not stop the computation.
-    Args:
-        src: The source of the video file (can be local or url)
-        env: the build environment
-    Returns:
-        the src file, the extention suffix
-    """
-    if not bool(urlparse(src).netloc):
-        env.images.add_file("", src)
-
-    suffix = Path(src).suffix
-    if suffix not in SUPPORTED_MIME_TYPES:
-        logger.warning(
-            'The provided file type ("{}") is not a supported format. defaulting to ""'.format(
-                suffix
-            )
-        )
-    type = SUPPORTED_MIME_TYPES.get(suffix, "")
-
-    return (src, type)
-
-
-class video_node(nodes.General, nodes.Element):
-    """
-    Video node combining nodes.General with nodes.Element.
-    Written lowercase to match html naming convention.
+    A docutils node corresponding to a source html element.
     """
 
     pass
 
 
-class Video(SphinxDirective):
-    """Video directive.
-    Wrapper for the html <video> tag embeding all the supported options
+class video(nodes.General, nodes.TextElement):
+    """
+    A docutils node corresponding to a video html element.
     """
 
-    has_content: bool = True
-    required_arguments: int = 1
-    optional_arguments: int = 1
-    option_spec: Dict[str, Any] = {
-        "alt": directives.unchanged,
-        "autoplay": directives.flag,
-        "nocontrols": directives.flag,
-        "height": directives.unchanged,
-        "loop": directives.flag,
-        "muted": directives.flag,
-        "poster": directives.unchanged,
-        "preload": directives.unchanged,
-        "width": directives.unchanged,
-    }
+    pass
 
-    def run(self) -> List[video_node]:
-        """Return the video node based on the set options."""
-        env: BuildEnvironment = self.env
 
-        # check options that need to be specific values
-        height: str = self.options.get("height", "")
-        if height and not height.isdigit():
-            logger.warning(
-                "The provided height ({}) is ignored as it is not an integer".format(
-                    height
-                )
-            )
-            height = ""
+"""
+The build process for images is as follows:
+1. A collector runs and does some image stuff. This adds images to the environment and fills in candidates, which
+    enables transforms later on.
 
-        width: str = self.options.get("width", "")
-        if width and not width.isdigit():
-            logger.warning(
-                "The provided width ({}) is ignored as it is not an integer".format(
-                    width
-                )
-            )
-            width = ""
+The sphinx image collector is here:
+https://github.com/sphinx-doc/sphinx/blob/ff852bc7c31f48e66100e4647749fc199d92ca79/sphinx/environment/collectors/asset.py
 
-        preload: str = self.options.get("preload", "auto")
-        valid_preload = ["auto", "metadata", "none"]
-        if preload not in valid_preload:
-            logger.warning(
-                'The provided preload ({}) is not an accepted value. defaulting to "auto"'.format(
-                    preload
-                )
-            )
-            preload = "auto"
+2. Image converters change files from one type to another. 
+    They do this using a combination of candidates and updates to the environment and uri to correct the paths.
+    We skip this step since we don't worry about changing one video to another.
+3. Images are moved from the environment to their specific folders by the appropriate builders. Note this process
+    is different for each builder.
+    The builder also updates uris to be in-place relative to the correct target folder.
+    We update the builder to implement this logic. However, this logic currently targets all https:// links, which 
+    is possibly a problem.
+"""
 
-        # add the primary video files as images in the builder
-        primary_src = get_video(self.arguments[0], env)
 
-        # add the secondary video files as images in the builder if necessary
-        secondary_src = None
-        if len(self.arguments) == 2:
-            secondary_src = get_video(self.arguments[1], env)
-        elif env.config.video_enforce_extra_source is True:
-            logger.warning(
-                'A secondary source should be provided for "{}"'.format(
-                    self.arguments[0]
-                )
-            )
+class VideoCollector(collectors.EnvironmentCollector):
+    def clear_doc(
+        self, app: application.Sphinx, env: environment.BuildEnvironment, docname: str
+    ) -> None:
+        env.images.purge_doc(docname)
 
-        return [
-            video_node(
-                primary_src=primary_src,
-                secondary_src=secondary_src,
-                alt=self.options.get("alt", ""),
-                autoplay="autoplay" in self.options,
-                controls="nocontrols" not in self.options,
-                height=height,
-                loop="loop" in self.options,
-                muted="muted" in self.options,
-                poster=self.options.get("poster", ""),
-                preload=preload,
-                width=width,
-            )
+    def merge_other(
+        self,
+        app: application.Sphinx,
+        env: environment.BuildEnvironment,
+        docnames: set[str],
+        other: environment.BuildEnvironment,
+    ) -> None:
+        env.images.merge_other(docnames, other.images)
+
+    def process_doc(self, app: application.Sphinx, doctree: nodes.document) -> None:
+        for node in list(doctree.findall(video)) + list(doctree.findall(source)):
+            docname = app.env.docname
+            image_uri, _ = app.env.relfn2path(node["src"], docname)
+            node["src"] = image_uri
+            app.env.dependencies[docname].add(image_uri)
+            app.env.images.add_file(docname, image_uri)
+
+
+class VideoBuilder(html_builders.StandaloneHTMLBuilder):
+    def post_process_images(self, doctree: nodes.Node) -> None:
+        super().post_process_images(doctree)
+        for node in list(doctree.findall(video)) + list(doctree.findall(source)):
+            self.images[node["src"]] = self.env.images[node["src"]][1]
+
+
+class VideoTranslator(html_writers.HTMLTranslator, docutils.SphinxTranslator):
+    def _get_src_path(self, src: str) -> str:
+        return str(pathlib.PurePath(self.builder.imgpath) / pathlib.PurePath(src).name)
+
+    def visit_source(self, node: source) -> None:
+        src_path = self._get_src_path(node["src"])
+        attributes = {"src": src_path, "type": node["type"]}
+        self.body.append(self.emptytag(node, "source", **attributes))
+
+    def depart_source(self, _: source) -> None:
+        """Exit the video node."""
+        pass
+
+    def visit_video(self, node: video) -> None:
+        if "src" in node:
+            node["src"] = self._get_src_path(node["src"])
+
+        # key value attributes
+        attributes: List[str] = [
+            '{k} = "{v}"'.format(k=k, v=node[k])
+            for k in [
+                "controlslist",
+                "crossorigin",
+                "height",
+                "width",
+                "poster",
+                "preload",
+                "src",
+            ]
+            if k in node
         ]
 
+        # boolean attributes
+        attributes.extend(
+            [
+                k
+                for k in [
+                    "autoplay",
+                    "autopictureinpicture",
+                    "controls",
+                    "disablepictureinpicture",
+                    "disableremoteplayback",
+                    "loop",
+                    "playsinline",
+                    "muted",
+                ]
+                if k in node and node[k]  # value is truthy
+            ]
+        )
+        self.body.append("<video {}>\n".format(" ".join(attributes)))
 
-def visit_video_node_html(translator: SphinxTranslator, node: video_node) -> None:
-    """Entry point of the html video node."""
-    # start the video block
-    attr: List[str] = [f'{k}="{node[k]}"' for k in SUPPORTED_OPTIONS if node[k]]
-    html: str = f"<video {' '.join(attr)}>"
-
-    # build the sources
-    html_source = '<source src="{}" type="{}">'
-    html += html_source.format(*node["primary_src"])
-    if node["secondary_src"] is not None:
-        html += html_source.format(*node["secondary_src"])
-
-    # add the alternative message
-    html += node["alt"]
-
-    translator.body.append(html)
+    def depart_video(self, _: video) -> None:
+        """Exit the video node."""
+        self.body.append("</video>\n")
 
 
-def depart_video_node_html(translator: SphinxTranslator, node: video_node) -> None:
-    """Exit of the html video node."""
-    translator.body.append("</video>")
-
-
-def visit_video_node_unsupported(
-    translator: SphinxTranslator, node: video_node
+# No args might not be valid here, I haven't checked
+def visit_node_unsupported(
+    translator: docutils.SphinxTranslator, node: nodes.Node
 ) -> None:
-    """Entry point of the ignored video node."""
-    logger.warning(
-        "video {}: unsupported output format (node skipped)".format(node["primary_src"])
-    )
     raise nodes.SkipNode
 
 
-def setup(app: Sphinx) -> Dict[str, bool]:
-    """Add video node and parameters to the Sphinx builder."""
-    app.add_config_value("video_enforce_extra_source", False, "html")
-    app.add_node(
-        video_node,
-        html=(visit_video_node_html, depart_video_node_html),
-        epub=(visit_video_node_unsupported, None),
-        latex=(visit_video_node_unsupported, None),
-        man=(visit_video_node_unsupported, None),
-        texinfo=(visit_video_node_unsupported, None),
-        text=(visit_video_node_unsupported, None),
-    )
-    app.add_directive("video", Video)
+unsupported_tuple = (visit_node_unsupported, None)
+unsupported_dict = dict(
+    [(k, unsupported_tuple) for k in ["epub", "latex", "man", "texinfo", "text"]]
+)
 
-    return {
-        "parallel_read_safe": True,
-        "parallel_write_safe": True,
-    }
+
+def register_video_nodes(app: application.Sphinx) -> None:
+    """
+    Registers video nodes and the updated translator with sphinx
+    """
+    app.add_env_collector(VideoCollector)
+    app.add_builder(VideoBuilder, override=True)
+    app.set_translator("html", VideoTranslator)
+
+    app.add_node(video, **unsupported_dict)
+    app.add_node(source, **unsupported_dict)
